chg: Move API into a new file, cleanup
